Use HOB as the Source of Memory Protection Settings
A HOB produced in PEI phase will now be the source of the memory protection settings for the current boot. PEI phase will now always have null pointer protection, stack guard, and stack no-execute which consequently requires page tables to always be built in PEI (happens on our platforms anyways).

This PR comments out the fixed at build PCDs which previously defined the memory protection settings to make it clear that they are deprecated. The HOB containing the memory protection settings are close to a 1:1 of the PCDs, but are instead implemented as bitfields. Changing to this HOB implementation allows memory protection settings to be configured each boot based on platform-relevant conditions.